+++ src/app/shared/scroll.model.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_SCROLL_OPTIONS: any = {behavior: 'smooth'};
