+++ src/index.html
@@ -9,6 +9,6 @@
   <link rel="icon" type="image/x-icon" href="favicon.ico">
 </head>
 <body>
-  <app></app>
+<lc-app></lc-app>
 </body>
 </html>
